Renamed & moved frontend router.
